Refactor plugin into smaller services; use actual types from Homebridge package
