+++ .vscode/settings.json
@@ -1,3 +1,9 @@
 {
-  "editor.formatOnSave": true
+  "editor.formatOnSave": true,
+  "[typescript]": {
+    "editor.codeActionsOnSave": [
+      "source.organizeImports"
+    ],
+  },
+  "typescript.tsserver.experimental.enableProjectDiagnostics": true
 }
