+++ .vscode/tasks.json
@@ -1,0 +1,20 @@
+{
+	"version": "2.0.0",
+	"tasks": [
+		{
+			"type": "npm",
+			"script": "build",
+			"group": "build",
+			"problemMatcher": [],
+			"label": "npm: build",
+			"detail": "rollup -c"
+		},
+		{
+			"type": "npm",
+			"script": "dev",
+			"group": "build",
+			"label": "npm: dev",
+			"detail": "rollup -c"
+		}
+	]
+}
