+++ config.schema.json
@@ -23,11 +23,12 @@
         "default": "",
         "description": "Read the <a href='https://github.com/nfarina/homebridge-tesla'>documentation</a> to learn how to get a refresh token."
       },
-      "password": {
-        "title": "Tesla password (only required for keyless driving)",
+      "prefix": {
+        "title": "Services Prefix",
         "type": "string",
         "required": false,
-        "default": ""
+        "default": "",
+        "description": "Prefix for all services. Useful if you have multiple Tesla vehicles. Example: 'Climate' becomes 'Model S Climate'."
       },
       "waitMinutes": {
         "title": "Time in minutes to wait for the car to wake up (default is 1 minute)",
@@ -42,63 +43,77 @@
         "title": "Latitude of HomeLink device",
         "type": "string",
         "required": false,
-        "description": "This is the latitude position of your HomeLink device."
+        "description": "This is the latitude position of your HomeLink device.",
+        "default": ""
       },
       "longitude": {
         "title": "Longitude of HomeLink device",
         "type": "string",
         "required": false,
-        "description": "This is the longitude position of your HomeLink device."
+        "description": "This is the longitude position of your HomeLink device.",
+        "default": ""
       },
       "disableDoors": {
         "title": "Disable Door Lock Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableSentryMode": {
         "title": "Disable Sentry Mode Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "sentryModeSwitch": {
         "title": "Sentry Mode Service as a Switch",
         "type": "boolean",
-        "description": "This will configure Sentry Mode as a switch, instead of a lock."
+        "description": "This will configure Sentry Mode as a switch, instead of a lock.",
+        "default": false
       },
       "disableClimate": {
         "title": "Disable Climate Control Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableDefrost": {
         "title": "Disable Defrost Control Service",
         "type": "boolean",
-        "description": "If climate control service is disabled, this switch will also be disabled."
+        "description": "If climate control service is disabled, this switch will also be disabled.",
+        "default": false
       },
       "disableTrunk": {
         "title": "Disable Trunk Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableFrunk": {
         "title": "Disable Frunk Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableCharger": {
         "title": "Disable Charger Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableChargePort": {
         "title": "Disable Charge Port Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableStarter": {
         "title": "Disable Starter Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "enableHomeLink": {
         "title": "Enable HomeLink Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       },
       "disableChargeLevel": {
         "title": "Disable Charge Level Service",
-        "type": "boolean"
+        "type": "boolean",
+        "default": false
       }
     }
   }
