+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "homebridge-tesla",
-  "version": "3.4.4",
+  "version": "4.0.0",
   "description": "Tesla support for Homebridge: https://github.com/nfarina/homebridge",
   "license": "ISC",
   "keywords": [
@@ -22,26 +22,27 @@
     "dist",
     "config.schema.json"
   ],
+  "scripts": {
+    "dev": "rollup -c -w --bundleConfigAsCjs",
+    "build": "rollup -c --bundleConfigAsCjs",
+    "prepublishOnly": "npm run build"
+  },
   "dependencies": {
     "@babel/polyfill": "^7.4.4",
     "promptly": "^3.2.0",
-    "teslajs": "^4.9.11"
+    "teslajs": "^4.10.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/plugin-proposal-class-properties": "^7.16.0",
-    "@babel/preset-env": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
+    "@babel/core": "^7.19.6",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.19.4",
+    "@babel/preset-typescript": "^7.18.6",
     "@types/promptly": "^3.0.2",
-    "prettier": "^2.4.1",
-    "rollup": "^2.59.0",
+    "homebridge": "^1.5.1",
+    "prettier": "^2.7.1",
+    "rollup": "^3.2.3",
     "rollup-plugin-babel": "^4.3.2",
     "rollup-plugin-node-resolve": "^5.0.2",
-    "typescript": "^4.4.4"
-  },
-  "scripts": {
-    "dev": "rollup -c -w",
-    "build": "rollup -c",
-    "prepublishOnly": "npm run build"
+    "typescript": "^4.8.4"
   }
 }
