+++ tsconfig.json
@@ -3,9 +3,11 @@
 		// Target latest version of ECMAScript.
 		"target": "esnext",
 		// Allow features like dynamic import().
-		"module": "esnext",
+		"module": "CommonJS",
 		// Search under node_modules for non-relative imports.
 		"moduleResolution": "node",
+		// Allow importing our config schema.
+		"resolveJsonModule": true,
 		// Process & infer types from .js files.
 		"allowJs": false,
 		// Don't emit; allow Babel to transform files.
@@ -13,13 +15,16 @@
 		// Enable strictest settings like strictNullChecks & noImplicitAny.
 		"strict": true,
 		// Disallow features that require cross-file information for emit.
-		"isolatedModules": true,
+		// Actually needed because Homebridge uses CONST ENUM.
+		// "isolatedModules": true,
 		// Allow importing non-TS files as "any".
 		"noImplicitAny": false,
 		// Import non-ES modules as default imports.
 		"esModuleInterop": true,
 		// Allow fancy language features.
-		"lib": ["esnext"]
+		"lib": [
+			"esnext"
+		]
 	},
 	"include": [
 		"src"
